+++ mobile_frontend/lib/models/activity.dart
