+++ mobile_frontend/lib/models/notif.dart
