+++ mobile_frontend/lib/screens/home/category_scroller.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_frontend/dart_src/activity.dart';
+import 'package:mobile_frontend/models/activity.dart';
 
 class CategoryScroller extends StatefulWidget {
   @override
