+++ mobile_frontend/lib/screens/home/profile.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:mobile_frontend/main_drawer.dart';
+import 'package:mobile_frontend/screens/home/main_drawer.dart';
 
 class Profile extends StatefulWidget {
   @override
